backend/requirements.txt
# UDHBHA Backend Dependencies
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
bcrypt>=4.1.0
PyJWT>=2.8.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pg8000>=1.30.0
httpx==0.24.1
httpcore==0.17.3
supabase==2.3.4
